requirements.txt
discord.py
discord.py[voice]
python-dotenv